+++ lib/webparts/itAssetManager/components/AssetAttachmentSection.d.ts
@@ -1,0 +1,9 @@
+import * as React from 'react';
+import { FileUploadService } from '../services/FileUploadService';
+interface IAssetAttachmentSectionProps {
+    assetId: string;
+    fileService: FileUploadService;
+}
+declare const AssetAttachmentSection: React.FC<IAssetAttachmentSectionProps>;
+export default AssetAttachmentSection;
+//# sourceMappingURL=AssetAttachmentSection.d.ts.map