+++ lib/webparts/itAssetManager/models/IAttachment.d.ts
@@ -1,0 +1,12 @@
+export declare type AttachmentCategory = 'purchase' | 'repairs' | 'gifted' | 'transfer' | 'scrap' | 'validation' | 'photos' | 'other';
+export interface IAttachment {
+    name: string;
+    serverRelativeUrl: string;
+    absoluteUrl: string;
+    downloadUrl: string;
+    category: AttachmentCategory;
+    timeCreated: string;
+    fileSize: number;
+    previewType: 'browser' | 'office' | 'download';
+}
+//# sourceMappingURL=IAttachment.d.ts.map